\name{input.data.TSPADIdata}
\alias{input.data.TSPADIdata}
\alias{input.periods.TSPADIdata}
\alias{output.data.TSPADIdata}
\alias{output.periods.TSPADIdata}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \method{input.data}{TSPADIdata}(x, series=seq(length=input.dimension(x)))
    \method{output.data}{TSPADIdata}(x,series=seq(length=output.dimension(x)))
    \method{input.periods}{TSPADIdata}(x)
    \method{output.periods}{TSPADIdata}(x)
}
%\keyword{DSE}
\keyword{ts}


