\name{periods.tfPADIdata}
\alias{periods.tfPADIdata}
\alias{start.tfPADIdata}
\alias{end.tfPADIdata}
\alias{frequency.tfPADIdata}
\alias{series.names.tfPADIdata}

\title{Specific Methods for Tframed data}
\description{See the generic function description.}
\usage{
    periods(x)
    start(x)
    end(x)
    frequency(x)
    freeze(x)
    series.names(x)
}

\seealso{
    \code{\link[tframe]{periods}}, 
    \code{\link[tframe]{start}},
    \code{\link[tframe]{end}},
    \code{\link[tframe]{frequency}},
    \code{\link[tframe]{freeze}},
    \code{\link[tframe]{series.names}}
}

%\keyword{DSE}
\keyword{ts}

