\name{source.info.TSPADIdata}
\alias{source.info.TSPADIdata}
\alias{sourcedb.TSPADIdata}
\alias{sourceserver.TSPADIdata}
\alias{availability.TSPADIdata}
\alias{identifiers.TSPADIdata}
\alias{periods.TSPADIdata}

\title{TSPADIdata Specific Methods}
\description{See the generic function description.}
\usage{
    \method{source.info}{TSPADIdata}(obj)
    \method{sourcedb}{TSPADIdata}(obj)
    \method{sourceserver}{TSPADIdata}(obj)
    \method{availability}{TSPADIdata}(obj, verbose=T, timeout=60)
    \method{identifiers}{TSPADIdata}(obj)
    \method{periods}{TSPADIdata}(x)
}
\seealso{
    \code{\link{source.info}} 
}
%\keyword{DSE}
\keyword{ts}

