\name{sourceInfo.TSdata}
\alias{sourceInfo.TSdata}
\alias{sourceInfo.TSestModel}
\alias{sourcedb.TSdata}
\alias{sourcedb.TSestModel}
\alias{sourceserver.TSdata}
\alias{sourceserver.TSestModel}
\alias{identifiers.TSdata}
\alias{identifiers.TSestModel}

\title{tfPADI Specific Methods}
\description{See the generic function description.}
\usage{
    \method{sourceInfo}{TSdata}(obj)
    \method{sourceInfo}{TSestModel}(obj)
    
    \method{sourcedb}{TSdata}(obj)
    \method{sourcedb}{TSestModel}(obj)
    
    \method{sourceserver}{TSdata}(obj)
    \method{sourceserver}{TSestModel}(obj)
    
    \method{identifiers}{TSdata}(obj)
    \method{identifiers}{TSestModel}(obj)    
}
\arguments{
    \item{obj}{An object of class TSdata or TSestModel.}
}

\seealso{ \code{\link[dsepadi]{sourceInfo}} }

\concept{DSE}
\keyword{ts}

