\name{tfperiods.tfPADIdata}
\alias{tframe.tfPADIdata}
\alias{tfperiods.tfPADIdata}
\alias{tfstart.tfPADIdata}
\alias{tfend.tfPADIdata}
\alias{tffrequency.tfPADIdata}
\alias{seriesNames.tfPADIdata}

\title{Specific Methods for Tframed data}
\description{See the generic function description.}
\usage{
    \method{tframe}{tfPADIdata}(x)
    \method{tfperiods}{tfPADIdata}(x)
    \method{tfstart}{tfPADIdata}(x, ...)
    \method{tfend}{tfPADIdata}(x, ...)
    \method{tffrequency}{tfPADIdata}(x, ...)
    \method{seriesNames}{tfPADIdata}(x)
}
\arguments{
    \item{x}{a tfPADIdata object.}
    \item{...}{(further arguments, currently disregarded).}
}
\seealso{
    \code{\link[tframe]{tframe}}, 
    \code{\link[tframe]{tfperiods}}, 
    \code{\link[tframe]{tfstart}},
    \code{\link[tframe]{tfend}},
    \code{\link[tframe]{tffrequency}},
    \code{\link[tframe]{freeze}},
    \code{\link[tframe]{seriesNames}}
}

\concept{DSE}
\keyword{ts}

