\name{input.data.TSPADIdata}
\alias{input.data.TSPADIdata}
\alias{input.periods.TSPADIdata}
\alias{output.data.TSPADIdata}
\alias{output.periods.TSPADIdata}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
	input.data(x, ...)
	output.data(x, ...)
	input.periods(x, ...)
	output.periods(x, ...)
}
\keyword{DSE}
\keyword{ts}


