\name{modify}
\alias{modify}
\title{modify a database descriptor object}
\usage{modify(obj, ...)}
\arguments{
    \item{obj}{database descriptor object to be modified.}
    \item{...}{arguments to be passed to other methods.}
}
\description{Generic method. See specific methods for details.}
\seealso{
\code{\link{modify.tfPADIdata}} 
\code{\link[dse2]{modify.TSPADIdata}} 
\code{\link[tframe]{freeze}}
}
%\keyword{tframe}
\keyword{ts}


