\name{modify}
\alias{modify}
\title{modify a database descriptor object}
\usage{modify(obj, ...)}
\description{Generic method. See specific methods for details.}
\seealso{
\code{\link{modify.tfPADIdata}} 
\code{\link[dse2]{modify.TSPADIdata}} 
\code{\link[tframe]{freeze}}
}
\keyword{tframe}
\keyword{time series data}


