\name{periods.tfPADIdata}
\alias{periods.tfPADIdata}
\alias{start.tfPADIdata}
\alias{end.tfPADIdata}
\alias{frequency.tfPADIdata}
\alias{series.names.tfPADIdata}

\title{Specific Methods for Tframed data}
\description{See the generic function description.}
\usage{
    \method{periods}{tfPADIdata}(x)
    \method{start}{tfPADIdata}(x)
    \method{end}{tfPADIdata}(x)
    \method{frequency}{tfPADIdata}(x)
    \method{series.names}{tfPADIdata}(x)
}
\arguments{
    \item{x}{a tfPADIdata object.}
}
\seealso{
    \code{\link[tframe]{periods}}, 
    \code{\link[tframe]{start}},
    \code{\link[tframe]{end}},
    \code{\link[tframe]{frequency}},
    \code{\link[tframe]{freeze}},
    \code{\link[tframe]{series.names}}
}

%\keyword{DSE}
\keyword{ts}

