\name{source.info.TSPADIdata}
\alias{source.info.TSPADIdata}
\alias{sourcedb.TSPADIdata}
\alias{sourceserver.TSPADIdata}
\alias{availability.TSPADIdata}
\alias{identifiers.TSPADIdata}
\alias{periods.TSPADIdata}

\title{TSPADIdata Specific Methods}
\description{See the generic function description.}
\usage{
	source.info(x)
	sourcedb(x)
	sourceserver(x)
	availability(x)
	identifiers(x)
	periods(x)
}
\seealso{
	\code{\link{source.info}} 
}
\keyword{DSE}
\keyword{ts}

